UI changes and types check
